VanDog/requirements.txt
Flask==2.0.1
numpy==1.18.5
requests==2.22.0
requests-oauthlib==1.3.0
tensorflow==1.15.0
keras==2.2.5
h5py==2.10.0
Werkzeug==2.0.1
setuptools==62.3.2
Cython==0.29.30
wheel==0.37.1
pycocotools==2.0.4
scikit-image==0.18.3
opencv-python==4.1.2.30
protobuf==3.17.3